Conflicts:
	qxrd.pro
	qxrd.spec.in
	source/qxrddebug.cpp
	source/qxrddebug.h
	source/qxrdexperiment.cpp
	source/qxrdserver.cpp
	source/source.pro

+++ source/qxrdglobalpreferencesdialog.h
@@ -21,8 +21,8 @@ protected:
   void changeEvent(QEvent *e);
 
 private:
-  void setupDebugWidgets(int dbg);
-  int readDebugWidgets();
+  void setupDebugWidgets(qint64 dbg);
+  qint64 readDebugWidgets();
 
 private:
   QxrdApplication                 *m_Application;
